Fix falling-blocks pitch mapping: prefer keyboard detector, reject bad calibrations
The falling-blocks detector had two pitch mapping bugs:
1. Octave error: block-based mapping with base_octave=2 placed notes ~2 octaves
   too low when identify_c_position found the wrong grid position.
2. Semitone error: calibration with high residuals shifted notes across pitch
   boundaries (e.g., C#5 → C5).

Fixes:
- Invert pitch map priority: try keyboard detector first (reliable octave
  assignment), use block positions for x-calibration, fall back to pure
  block-based mapping only when keyboard detection fails.
- Reject calibration when max_residual > min_gap/2 (prevents pitch flips).
- Fix argparse --detector default ('falling-blocks' → None) so settings.json
  values are not shadowed by the argparse default.

Verified: test3 E6/D#6/B5 correct (was D4/B3/G#3), test4 C#5 correct (was C5).
All regression tests pass (15 passed, 1 xfailed).

+++ piano_video_scribe/config.py
@@ -95,7 +95,7 @@ Examples:
     p.add_argument('output', help='Path to output MIDI')
     # All settings-overridable args default to None so --settings can fill them in.
     # Hardcoded defaults are applied after settings are merged.
-    p.add_argument('--bpm', type=int, default=None,
+    p.add_argument('--bpm', type=float, default=None,
                    help='BPM of the song. Auto-detected from video audio if omitted.')
     p.add_argument('--frame', type=int, default=None,
                    help='Frame index for keyboard detection (default: 5, should be note-free)')
